Refactored Gallery Dicom Viewer
